Separate directory management for PG and RDF
