--- neptune-export/src/main/java/com/amazonaws/services/neptune/cli/CommonFileSystemModule.java
@@ -1,41 +1,0 @@
-/*
-Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-Licensed under the Apache License, Version 2.0 (the "License").
-You may not use this file except in compliance with the License.
-A copy of the License is located at
-    http://www.apache.org/licenses/LICENSE-2.0
-or in the "license" file accompanying this file. This file is distributed
-on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
-express or implied. See the License for the specific language governing
-permissions and limitations under the License.
-*/
-
-package com.amazonaws.services.neptune.cli;
-
-import com.amazonaws.services.neptune.io.Directories;
-import com.amazonaws.services.neptune.io.DirectoryStructure;
-import com.github.rvesse.airline.annotations.Option;
-import com.github.rvesse.airline.annotations.restrictions.Once;
-import com.github.rvesse.airline.annotations.restrictions.Path;
-import com.github.rvesse.airline.annotations.restrictions.PathKind;
-import com.github.rvesse.airline.annotations.restrictions.Required;
-
-import java.io.File;
-import java.io.IOException;
-
-public class CommonFileSystemModule {
-
-    @Option(name = {"-d", "--dir"}, description = "Root directory for output")
-    @Required
-    @Path(mustExist = false, kind = PathKind.DIRECTORY)
-    @Once
-    private File directory;
-
-    @Option(name = {"-t", "--tag"}, description = "Directory prefix (optional)")
-    @Once
-    private String tag = "";
-
-    public Directories createDirectories(DirectoryStructure directoryStructure) throws IOException {
-        return Directories.createFor(directoryStructure, directory, tag );
-    }
-}
